refactor: refactor elasticsearch api fetch implementation
