+++ src/datasources/index.ts
@@ -3,3 +3,4 @@ export * from './sourceFileApi';
 export * from './fetchApi';
 export * from './chatBotApi';
 export * from './dynamoApi.ts';
+export * from './esApi.ts';
