+++ src/store/clusterManageStore.ts
@@ -1,56 +1,8 @@
 import { defineStore } from 'pinia';
-import { loadHttpClient } from '../datasources';
+import { ClusterAlias, ClusterIndex, ClusterNode, esApi, loadHttpClient } from '../datasources';
 import { lang } from '../lang';
 import { Connection, DatabaseType } from './connectionStore.ts';
-import { CustomError, debug, jsonify, optionalToNullableInt } from '../common';
-
-export enum IndexHealth {
-  GREEN = 'green',
-  YELLOW = 'yellow',
-  RED = 'red',
-}
-
-export enum IndexStatus {
-  OPEN = 'open',
-  CLOSED = 'closed',
-  HIDDEN = 'hidden',
-}
-
-export type ClusterIndex = {
-  index: string;
-  uuid: string;
-  health: IndexHealth;
-  status: IndexStatus;
-  storage: string;
-  shards: {
-    primary: number;
-    replica: number;
-  };
-  docs: {
-    count: number;
-    deleted: number;
-  };
-};
-export type ClusterAlias = {
-  alias: string;
-  index: string;
-  filter: string;
-  routing: {
-    index: string;
-    search: string;
-  };
-  isWriteIndex: boolean;
-};
-
-export enum NodeRoleEnum {
-  DATA = 'DATA',
-  INGEST = 'INGEST',
-  MASTER_ELIGIBLE = 'MASTER_ELIGIBLE',
-  ML = 'ML',
-  REMOTE_CLUSTER_CLIENT = 'REMOTE_CLUSTER_CLIENT',
-  TRANSFORM = 'TRANSFORM',
-  COORDINATING = 'COORDINATING',
-}
+import { CustomError, debug, optionalToNullableInt } from '../common';
 
 export enum ShardStateEnum {
   UNASSIGNED = 'UNASSIGNED',
@@ -175,19 +127,6 @@ export type ShardState = {
     reason: string;
   };
 };
-export type SearchNode = {
-  ip: string;
-  name: string;
-  roles: Array<string>;
-  master: string;
-  heap: {
-    percent: string;
-  };
-  ram: {
-    percent: string;
-  };
-  cpu: string;
-};
 
 export enum TemplateType {
   INDEX_TEMPLATE = 'INDEX_TEMPLATE',
@@ -223,7 +162,7 @@ export type RawClusterStats = {
       master: number;
       data: number;
     };
-    instances: Array<SearchNode>;
+    instances: Array<ClusterNode>;
     versions: Array<string>;
   };
   indices: {
@@ -247,7 +186,7 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
   state: (): {
     connection: Connection | undefined;
     cluster: RawClusterStats | undefined;
-    nodes: Array<SearchNode>;
+    nodes: Array<ClusterNode>;
     shards: Array<Shard>;
     indices: Array<ClusterIndex>;
     aliases: Array<ClusterAlias>;
@@ -308,121 +247,7 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
     async fetchNodes() {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        try {
-          const data = await client.get('/_cat/nodes', 'format=json');
-          this.nodes = data
-            .map(
-              (node: {
-                ip: string;
-                name: string;
-                'node.role': string;
-                master: string;
-                'heap.percent': string;
-                'ram.percent': string;
-                cpu: string;
-              }) => ({
-                ip: node.ip,
-                name: node.name,
-                roles: node['node.role']
-                  .split('')
-                  .map((char: string) => {
-                    switch (char) {
-                      case 'd':
-                        return NodeRoleEnum.DATA;
-                      case 'i':
-                        return NodeRoleEnum.INGEST;
-                      case 'm':
-                        return NodeRoleEnum.MASTER_ELIGIBLE;
-                      case 'l':
-                        return NodeRoleEnum.ML;
-                      case 'r':
-                        return NodeRoleEnum.REMOTE_CLUSTER_CLIENT;
-                      case 't':
-                        return NodeRoleEnum.TRANSFORM;
-                      case '-':
-                        return NodeRoleEnum.COORDINATING;
-                      default:
-                        return '';
-                    }
-                  })
-                  .filter((role: string) => role !== ''),
-                master: node.master === '*',
-                heap: {
-                  percent: node['heap.percent'],
-                },
-                ram: {
-                  percent: node['ram.percent'],
-                },
-                cpu: node.cpu,
-              }),
-            )
-            .sort((a: SearchNode, b: SearchNode) => a.name.localeCompare(b.name));
-        } catch (err) {
-          debug(`Failed to fetch nodes: ${err}`);
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
-      } else {
-        this.nodes = [];
-      }
-    },
-    async fetchNodeState(nodeName: string) {
-      if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
-      if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        const data = await client.get(
-          `/_cat/nodes`,
-          'format=json&bytes=b&h=ip,id,name,heap.percent,heap.current,heap.max,ram.percent,ram.current,ram.max,node.role,master,cpu,load_1m,load_5m,load_15m,disk.used_percent,disk.used,disk.total,shard_stats.total_count,mappings.total_count&full_id=true',
-        );
-        return Object.entries<{
-          ip: string;
-          name: string;
-          version: string;
-          'heap.percent': string;
-          'heap.current': string;
-          'heap.max': string;
-          'ram.percent': string;
-          'ram.current': string;
-          'ram.max': string;
-          'disk.used_percent': string;
-          'disk.used': string;
-          'disk.total': string;
-          'shard_stats.total_count': string;
-          'mappings.total_count': string;
-          cpu: string;
-        }>(data)
-          .map(([id, node]) => ({
-            id,
-            version: node.version,
-            ip: node.ip,
-            cpu: node.cpu,
-            name: node.name,
-            heap: {
-              percent: parseInt(node['heap.percent']),
-              current: parseInt(node['heap.current']),
-              max: parseInt(node['heap.max']),
-            },
-            ram: {
-              percent: parseInt(node['ram.percent']),
-              current: parseInt(node['ram.current']),
-              max: parseInt(node['ram.max']),
-            },
-            disk: {
-              percent: parseInt(node['disk.used_percent']),
-              current: parseInt(node['disk.used']),
-              max: parseInt(node['disk.total']),
-            },
-            shard: {
-              total: parseInt(node['shard_stats.total_count']),
-            },
-            mapping: {
-              total: parseInt(node['mappings.total_count']),
-            },
-          }))
-          .find(({ name }) => name === nodeName);
+        this.nodes = await esApi.catNodes(this.connection);
       } else {
         this.nodes = [];
       }
@@ -640,30 +465,11 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
     async fetchIndices() {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        const data = (await client.get('/_cat/indices', 'format=json&s=index')) as Array<{
-          [key: string]: string;
-        }>;
+        const indices = await esApi.catIndices(this.connection);
 
-        this.indices = data
-          .filter((index: { [key: string]: string }) =>
-            this.hideSystemIndices ? !index.index.startsWith('.') : true,
-          )
-          .map((index: { [key: string]: string }) => ({
-            index: index.index,
-            uuid: index.uuid,
-            health: index.health as IndexHealth,
-            status: index.status as IndexStatus,
-            storage: index['store.size'],
-            shards: {
-              primary: parseInt(index['pri'], 10),
-              replica: parseInt(index['rep'], 10),
-            },
-            docs: {
-              count: parseInt(index['docs.count'] || '0', 10),
-              deleted: parseInt(index['docs.deleted'], 10),
-            },
-          }));
+        this.indices = indices.filter(index =>
+          this.hideSystemIndices ? !index.index.startsWith('.') : true,
+        );
       } else {
         this.indices = [];
       }
@@ -671,24 +477,11 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
     async fetchAliases() {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        const data = (await client.get('/_cat/aliases', 'format=json&s=alias')) as Array<{
-          [key: string]: string;
-        }>;
-        this.aliases = data
-          .filter((alias: { [key: string]: string }) =>
-            this.hideSystemIndices ? !alias.index.startsWith('.') : true,
-          )
-          .map((alias: { [key: string]: string }) => ({
-            alias: alias.alias,
-            index: alias.index,
-            filter: alias.filter,
-            routing: {
-              index: alias['routing.index'],
-              search: alias['routing.search'],
-            },
-            isWriteIndex: alias['is_write_index'] === 'true',
-          }));
+        const aliases = await esApi.catAliases(this.connection);
+
+        this.aliases = aliases.filter(alias =>
+          this.hideSystemIndices ? !alias.index.startsWith('.') : true,
+        );
       } else {
         this.aliases = [];
       }
@@ -739,15 +532,7 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
         this.templates = [];
       }
     },
-    async createIndex({
-      indexName,
-      shards,
-      replicas,
-      master_timeout,
-      timeout,
-      wait_for_active_shards,
-      body,
-    }: {
+    async createIndex(options: {
       indexName: string;
       shards?: number | null;
       replicas?: number | null;
@@ -758,69 +543,11 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
     }) {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-
-        const queryParams = new URLSearchParams();
-        [
-          { key: 'master_timeout', value: master_timeout },
-          { key: 'wait_for_active_shards', value: wait_for_active_shards },
-          { key: 'timeout', value: timeout },
-        ].forEach(param => {
-          if (param.value !== null && param.value !== undefined) {
-            queryParams.append(
-              param.key,
-              param.value.toString() + (param.key !== 'wait_for_active_shards' ? 's' : ''),
-            );
-          }
-        });
-
-        const parsedBody = body ? jsonify.parse(body) : undefined;
-
-        const payload =
-          parsedBody || shards || replicas
-            ? jsonify.stringify({
-                ...parsedBody,
-                settings: {
-                  number_of_shards: shards,
-                  number_of_replicas: replicas,
-                  ...(parsedBody?.settings || {}),
-                },
-              })
-            : undefined;
-
-        try {
-          const response = await client.put(
-            `/${indexName}`,
-            queryParams.toString() ?? undefined,
-            payload,
-          );
-
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.createIndex(this.connection, options);
       }
       return undefined;
     },
-    async createAlias({
-      aliasName,
-      indexName,
-      master_timeout,
-      timeout,
-      is_write_index,
-      filter,
-      routing,
-      search_routing,
-      index_routing,
-    }: {
+    async createAlias(options: {
       aliasName: string;
       indexName: string;
       master_timeout: number | null;
@@ -833,66 +560,13 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
     }) {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-
-        const queryParams = new URLSearchParams();
-        [
-          { key: 'master_timeout', value: master_timeout },
-          { key: 'timeout', value: timeout },
-        ].forEach(param => {
-          if (param.value !== null && param.value !== undefined) {
-            queryParams.append(param.key, param.value.toString() + 's');
-          }
-        });
-        const payload = {
-          actions: [
-            {
-              add: {
-                index: indexName,
-                alias: aliasName,
-                ...(is_write_index !== null && is_write_index !== undefined
-                  ? { is_write_index }
-                  : {}),
-                ...(filter ? { filter } : {}),
-                ...(routing ? { routing: `${routing}` } : {}),
-                ...(search_routing ? { search_routing: `${search_routing}` } : {}),
-                ...(index_routing ? { index_routing: `${index_routing}` } : {}),
-              },
-            },
-          ],
-        };
-
-        try {
-          const response = await client.post(
-            `/_aliases`,
-            queryParams.toString() ?? undefined,
-            jsonify.stringify(payload),
-          );
-
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.createAlias(this.connection, options);
         // refresh data
         Promise.all([this.fetchIndices(), this.fetchAliases()]).catch();
       }
       return undefined;
     },
-    async createTemplate({
-      name,
-      type,
-      create,
-      master_timeout,
-      body,
-    }: {
+    async createTemplate(options: {
       name: string;
       type: string;
       create?: boolean | null;
@@ -901,38 +575,7 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
     }) {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        const queryParams = new URLSearchParams();
-        [
-          { key: 'master_timeout', value: master_timeout },
-          { key: 'create', value: create },
-        ].forEach(param => {
-          if (param.value !== null && param.value !== undefined) {
-            queryParams.append(
-              param.key,
-              param.value.toString() + (param.key !== 'create' ? 's' : ''),
-            );
-          }
-        });
-
-        try {
-          const response = await client.put(
-            `/${type}/${name}`,
-            queryParams.toString(),
-            body ?? undefined,
-          );
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.createTemplate(this.connection, options);
       }
       return undefined;
     },
@@ -941,111 +584,35 @@ export const useClusterManageStore = defineStore('clusterManageStore', {
       // delete index
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        try {
-          const response = await client.delete(`/${indexName}`);
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.deleteIndex(this.connection, indexName);
       }
       return undefined;
     },
     async closeIndex(indexName: string) {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        try {
-          const response = await client.post(`/${indexName}/_close`);
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.closeIndex(this.connection, indexName);
       }
       return undefined;
     },
     async openIndex(indexName: string) {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        try {
-          const response = await client.post(`/${indexName}/_open`);
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.openIndex(this.connection, indexName);
       }
       return undefined;
     },
     async removeAlias(indexName: string, aliasName: string) {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        try {
-          const response = await client.delete(`/${indexName}/_alias/${aliasName}`);
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.removeAlias(this.connection, indexName, aliasName);
       }
       return undefined;
     },
     async switchAlias(aliasName: string, sourceIndexName: string, targetIndexName: string) {
       if (!this.connection) throw new Error(lang.global.t('connection.selectConnection'));
       if (this.connection.type === DatabaseType.ELASTICSEARCH) {
-        const client = loadHttpClient(this.connection);
-        const payload = {
-          actions: [
-            { remove: { index: sourceIndexName, alias: aliasName } },
-            { add: { index: targetIndexName, alias: aliasName } },
-          ],
-        };
-        try {
-          const response = await client.post(`/_aliases`, undefined, jsonify.stringify(payload));
-          if (response.status >= 300) {
-            throw new CustomError(
-              response.status,
-              `${response.error.type}: ${response.error.reason}`,
-            );
-          }
-        } catch (err) {
-          throw new CustomError(
-            err instanceof CustomError ? err.status : 500,
-            err instanceof CustomError ? err.details : (err as Error).message,
-          );
-        }
+        await esApi.switchAlias(this.connection, { aliasName, sourceIndexName, targetIndexName });
       }
       return undefined;
     },
